requirements.txt
sc2reader~=1.8.0
python-dotenv~=1.0.0